refactoring: move preparedDate to file
